mcp-servers/test-strategy-server/requirements.txt
# FastAPI for MCP server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Google Gemini AI
google-generativeai>=0.3.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
